Make the SimpleDateFormat parse dates expressed in UTC-00:00:00.
The previous fix that removes the Z is commit 0e282651. I still don’t
get why nor what it was trying to fix. AFAIK, when a PARSER receives an
input string that contains a Z (i.e. stands for Zulu time), it means
that it HAS to interpret the time as UTC-00:00:00. If the Z is not
present, we can make no assumption on how the parser will interpret the
time and it’s very likely that it will try to use its own locale to
deduce the timezone to resolve.

LLC-170